improved updateRoomDocument method and related classes. The constructor of RoomFields now takes an array of Fields[] and automatically converts them to an array of Maps each one corresponding to a set of fields.

+++ source/src/main/java/com/texnologia_logismikou/Cinematrix/DocumentObjects/Fields/ScreeningFields.java
@@ -6,7 +6,7 @@ public class ScreeningFields extends Fields {
 	private DoubleField time;
 	private DoubleField date;
 	
-	public ScreeningFields(String movie, Double time, Double date) {
+	public ScreeningFields(String movie, int time, long date) {
 		
 		this.movie = new StringField(movie);
 		this.time = new DoubleField(time);
